+++ plugins/languages.js
@@ -1,0 +1,3 @@
+export const lngDictionary = { current: {} };
+
+export const allLngValue = "--all--";