+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [0.9.0]
+### Added
+* DeepL configuration #27326
+* sync button with default for relation field #27326
+
+### Changed
+* improve ui/ux #27326
+
 ## [0.8.5]
 ### Added
 * RTL support for arabic language #27338